[sdk]: restructure IsmpClient module + refresh docs
