+++ docs/content/developers/evm/dispatching/post-responses.mdx
@@ -329,6 +329,6 @@ To learn how to self-relay responses:
 
 1. **[Get Started with the SDK](/developers/sdk/getting-started)** - Install and set up the Hyperbridge SDK
 2. **[Track POST Requests](/developers/sdk/tracking/post-requests)** - Monitor response status and extract delivery calldata
-3. **[IndexerClient API](/developers/sdk/api/indexer-client)** - Use the `postRequestStatusStream()` method to get calldata when response is ready for delivery
+3. **[IsmpClient API](/developers/sdk/api/ismp-client)** - Use the `postRequestStatusStream()` method to get calldata when response is ready for delivery
 
 The SDK automatically generates the proof and calldata you need to call the handler contract's `handlePostResponses` function on the destination chain. You can make your users self-relay responses or run a server which can relay responses on behalf of users.
