+++ docs/content/developers/sdk/api/evm-chain.mdx
@@ -566,4 +566,4 @@ interface GetProofParameters {
 ## Related
 
 - [SubstrateChain](/developers/sdk/api/substrate-chain) - For Substrate-based chains
-- [IndexerClient](/developers/sdk/api/indexer-client) - For querying cross-chain request status
+- [IsmpClient](/developers/sdk/api/ismp-client) - For querying cross-chain request status
