+++ docs/content/developers/sdk/api/substrate-chain.mdx
@@ -540,4 +540,4 @@ interface IIsmpMessage {
 ## Related
 
 - [EvmChain](/developers/sdk/api/evm-chain) - For EVM-compatible chains
-- [IndexerClient](/developers/sdk/api/indexer-client) - For querying cross-chain request status
+- [IsmpClient](/developers/sdk/api/ismp-client) - For querying cross-chain request status
