+++ docs/content/developers/sdk/api/token-gateway.mdx
@@ -1,0 +1,203 @@
+---
+title: TokenGateway
+description: API reference for TokenGateway — quote cross-chain token teleports and track their status via the Hyperbridge indexer.
+---
+
+# TokenGateway
+
+The `TokenGateway` class wraps the on-chain TokenGateway contract and provides
+utilities for estimating the native-token cost of a cross-chain teleport, plus
+(when configured with an indexer) snapshot queries and live streams of a
+teleport's status.
+
+Supports both EVM and Substrate chains as destination.
+
+## Constructor
+
+### new TokenGateway(params)
+
+```typescript lineNumbers
+new TokenGateway(params: { source: EvmChain; dest: EvmChain | SubstrateChain })
+```
+
+**Parameters:**
+
+| Name | Type | Description |
+|------|------|-------------|
+| `params.source` | `EvmChain` | Source EVM chain the teleport is dispatched from |
+| `params.dest` | `EvmChain \| SubstrateChain` | Destination chain the asset is delivered to |
+
+**Example:**
+
+```typescript lineNumbers
+import { TokenGateway, EvmChain, SubstrateChain } from "@hyperbridge/sdk"
+
+const source = EvmChain.fromParams({
+  chainId: 97,
+  rpcUrl: "https://data-seed-prebsc-1-s1.binance.org:8545",
+  host: "0x...",
+  consensusStateId: "BSC0",
+})
+const dest = EvmChain.fromParams({
+  chainId: 10200,
+  rpcUrl: "https://rpc.chiadochain.net",
+  host: "0x...",
+  consensusStateId: "GNO0",
+})
+
+const tokenGateway = new TokenGateway({ source, dest })
+```
+
+## Methods
+
+### quoteNative(params)
+
+Estimates the native-token cost of a teleport: relayer fee (for EVM
+destinations — automatically estimated from gas on the destination chain plus
+a 1% buffer) plus the Hyperbridge protocol fee (with a 1% buffer). Also returns
+the relayer fee converted into the source chain's fee-token denomination via
+the source-chain Uniswap V2 router.
+
+Returns `0` relayer fee for non-EVM destinations.
+
+```typescript lineNumbers
+async quoteNative(params: TeleportParams): Promise<QuoteNativeResult>
+```
+
+**Parameters:**
+
+| Name | Type | Description |
+|------|------|-------------|
+| `params.amount` | `bigint` | Amount to teleport |
+| `params.assetId` | `HexString` | Asset identifier (e.g. `keccak256(toHex("USDC"))`) |
+| `params.redeem` | `boolean` | Whether to redeem the underlying ERC-20 on the destination |
+| `params.to` | `HexString` | Recipient address (32-byte) |
+| `params.dest` | `string \| Uint8Array` | Destination state-machine ID |
+| `params.timeout` | `bigint` | Request timeout in seconds |
+| `params.data` | `HexString \| Uint8Array` | Optional destination contract calldata |
+
+**Returns:** `Promise<QuoteNativeResult>`
+
+| Field | Type | Description |
+|-------|------|-------------|
+| `totalNativeCost` | `bigint` | Total native-token cost (relayer + protocol fee, with 1% buffer) |
+| `relayerFeeInSourceFeeToken` | `bigint` | Relayer fee converted to the source chain's fee token |
+
+---
+
+### withQueryClient(queryClient, options?)
+
+Attaches a Hyperbridge indexer GraphQL client to this `TokenGateway` so that
+[`queryAssetTeleported`](#queryassetteleportedcommitment) and
+[`assetTeleportedStatusStream`](#assetteleportedstatusstreamcommitment) become
+available. Returns `this` for chaining.
+
+```typescript lineNumbers
+withQueryClient(
+  queryClient: IndexerQueryClient,
+  options?: { pollInterval?: number; tracing?: boolean },
+): this
+```
+
+**Parameters:**
+
+| Name | Type | Description |
+|------|------|-------------|
+| `queryClient` | `IndexerQueryClient` | GraphQL client from `createQueryClient(...)` |
+| `options.pollInterval` | `number` | Poll interval in milliseconds (default: `1_000`) |
+| `options.tracing` | `boolean` | Enable trace-level logging (default: `false`) |
+
+**Example:**
+
+```typescript lineNumbers
+import { createQueryClient } from "@hyperbridge/sdk"
+
+const queryClient = createQueryClient({ url: "https://indexer-url" })
+const tokenGateway = new TokenGateway({ source, dest }).withQueryClient(queryClient)
+```
+
+---
+
+### queryAssetTeleported(commitment)
+
+Queries an asset-teleported event by commitment hash from the Hyperbridge
+indexer.
+
+Requires a prior call to [`withQueryClient`](#withqueryclientqueryclient-options).
+
+```typescript lineNumbers
+async queryAssetTeleported(commitment: HexString): Promise<TokenGatewayAssetTeleportedWithStatus | undefined>
+```
+
+---
+
+### assetTeleportedStatusStream(commitment)
+
+Streams status updates for an asset-teleported event until it reaches a
+terminal state (`RECEIVED` or `REFUNDED`).
+
+Requires a prior call to [`withQueryClient`](#withqueryclientqueryclient-options).
+
+```typescript lineNumbers
+async *assetTeleportedStatusStream(commitment: HexString): AsyncGenerator<
+  {
+    status: TeleportStatus
+    metadata: {
+      blockHash: string
+      blockNumber: number
+      transactionHash: string
+      timestamp: bigint
+    }
+  },
+  void
+>
+```
+
+**Yields:**
+
+| Status | Description |
+|--------|-------------|
+| `TELEPORTED` | Teleport dispatched on the source chain |
+| `RECEIVED` | Asset delivered on the destination chain (terminal) |
+| `REFUNDED` | Teleport timed out and was refunded (terminal) |
+
+**Example:**
+
+```typescript lineNumbers
+for await (const update of tokenGateway.assetTeleportedStatusStream(commitment)) {
+  console.log(`Teleport status: ${update.status}`)
+  if (["RECEIVED", "REFUNDED"].includes(update.status)) break
+}
+```
+
+## Types
+
+### TeleportParams
+
+```typescript lineNumbers
+interface TeleportParams {
+  amount: bigint
+  assetId: HexString
+  redeem: boolean
+  to: HexString
+  dest: string | Uint8Array
+  timeout: bigint
+  data?: HexString | Uint8Array
+}
+```
+
+---
+
+### QuoteNativeResult
+
+```typescript lineNumbers
+interface QuoteNativeResult {
+  totalNativeCost: bigint
+  relayerFeeInSourceFeeToken: bigint
+}
+```
+
+## Related
+
+- [IsmpClient](/developers/sdk/api/ismp-client) — Track the underlying ISMP post-request that carries the teleport
+- [Getting Started](/developers/sdk/getting-started) — Initialise chain clients and the indexer
