+++ docs/content/developers/sdk/overview.mdx
@@ -13,7 +13,7 @@ The SDK consists of several key components that work together to provide a seaml
 
 | Component | Description |
 |-----------|-------------|
-| **IndexerClient** | Main client for querying the Hyperbridge indexer and monitoring request statuses |
+| **IsmpClient** | Main client for querying the Hyperbridge indexer and monitoring request statuses |
 | **EvmChain** | Implementation for EVM-compatible chains |
 | **SubstrateChain** | Implementation for Substrate-based chains |
