--- sdk/packages/sdk/src/client/GetRequestStatus.ts
@@ -1,246 +1,0 @@
-import { maxBy } from "lodash-es"
-
-// @ts-ignore
-import mergeRace from "@async-generator/merge-race"
-
-import type { IChain } from "@/chain"
-import {
-	type GetRequestWithStatus,
-	type HexString,
-	type RequestStatusKey,
-	type RequestStatusWithMetadata,
-	RequestStatus,
-} from "@/types"
-import { COMBINED_STATUS_WEIGHTS, REQUEST_STATUS_WEIGHTS } from "@/utils"
-import { AbortSignalInternal } from "@/utils/exceptions"
-
-import type { ProofFinalizer } from "./ProofFinalizer"
-import type { StateMachineQueries } from "./StateMachineQueries"
-import type { ClientContext } from "./types"
-import { waitOrAbort } from "./utils"
-
-/**
- * GET request status tracking — snapshot + streaming flows. Responses travel
- * back to the origin chain, so finality work happens against `source` rather
- * than `dest`.
- */
-export class GetRequestStatus {
-	private readonly logger
-
-	constructor(
-		private readonly ctx: ClientContext,
-		private readonly queries: StateMachineQueries,
-		private readonly proofFinalizer: ProofFinalizer,
-		private readonly timeoutStream: (
-			timeoutTimestamp: bigint,
-			chain: IChain,
-		) => AsyncGenerator<RequestStatusWithMetadata, void>,
-	) {
-		this.logger = ctx.logger.withTag("[GetRequestStatus]")
-	}
-
-	/**
-	 * Enhances a GET request with finality events (`SOURCE_FINALIZED`,
-	 * `HYPERBRIDGE_FINALIZED`).
-	 */
-	async addGetRequestFinalityEvents(request: GetRequestWithStatus): Promise<GetRequestWithStatus> {
-		const events: RequestStatusWithMetadata[] = []
-
-		const commit = () => {
-			this.logger.trace(`Added ${events.length} 'GetRequest' finality events`, events)
-			request.statuses = [...request.statuses, ...events]
-			return request
-		}
-
-		let hyperbridgeDelivered: RequestStatusWithMetadata | undefined
-
-		if (request.source === this.ctx.config.hyperbridge.config.stateMachineId) {
-			hyperbridgeDelivered = request.statuses[0]
-			return commit()
-		}
-
-		const sourceFinality = await this.queries.queryStateMachineUpdateByHeight({
-			statemachineId: request.source,
-			height: request.statuses[0].metadata.blockNumber,
-			chain: this.ctx.config.hyperbridge.config.stateMachineId,
-		})
-		if (!sourceFinality) return commit()
-
-		events.push({
-			status: RequestStatus.SOURCE_FINALIZED,
-			metadata: {
-				blockHash: sourceFinality.blockHash,
-				blockNumber: sourceFinality.height,
-				transactionHash: sourceFinality.transactionHash,
-				timestamp: sourceFinality.timestamp,
-			},
-		})
-
-		hyperbridgeDelivered = request.statuses.find((item) => item.status === RequestStatus.HYPERBRIDGE_DELIVERED)
-		if (!hyperbridgeDelivered) return commit()
-
-		try {
-			const response = await this.queries.queryResponseByRequestId(request.commitment)
-			if (!response) return commit()
-
-			const finalized = await this.proofFinalizer.buildGetResponseFinalized(
-				request,
-				hyperbridgeDelivered,
-				response,
-			)
-			if (finalized) events.push(finalized)
-		} catch (error) {
-			this.logger.trace("Could not generate HYPERBRIDGE_FINALIZED event for GET request:", error)
-		}
-
-		return commit()
-	}
-
-	/**
-	 * Snapshot query with all inferred finality events sorted.
-	 */
-	async queryGetRequestWithStatus(hash: HexString): Promise<GetRequestWithStatus | undefined> {
-		let request = await this.queries.queryGetRequest(hash)
-		if (!request) return
-
-		request = await this.addGetRequestFinalityEvents(request)
-
-		request.statuses = request.statuses.sort(
-			(a, b) => COMBINED_STATUS_WEIGHTS[a.status] - COMBINED_STATUS_WEIGHTS[b.status],
-		)
-		return request
-	}
-
-	/**
-	 * Streaming status updates for a GET request. Ends when the request reaches
-	 * the destination or a timeout becomes pending.
-	 */
-	async *getRequestStatusStream(hash: HexString): AsyncGenerator<RequestStatusWithMetadata, void> {
-		const controller = new AbortController()
-		try {
-			const request = await waitOrAbort(this.ctx, {
-				signal: controller.signal,
-				promise: () => this.queries.queryGetRequest(hash),
-			})
-
-			const chain = this.ctx.config.dest
-			const timeoutStream =
-				request.timeoutTimestamp > 0n ? this.timeoutStream(request.timeoutTimestamp, chain) : undefined
-			const statusStream = this.streamInternal(hash, controller.signal)
-			const combined = timeoutStream ? mergeRace(timeoutStream, statusStream) : statusStream
-
-			let item = await combined.next()
-			while (!item.done) {
-				yield item.value
-				item = await combined.next()
-			}
-		} catch (error) {
-			if (!AbortSignalInternal.isError(error)) throw error
-		}
-		controller.abort()
-	}
-
-	private async *streamInternal(
-		hash: HexString,
-		signal: AbortSignal,
-	): AsyncGenerator<RequestStatusWithMetadata, void> {
-		let request = await waitOrAbort(this.ctx, { signal, promise: () => this.queries.queryGetRequest(hash) })
-
-		let status: RequestStatusKey | undefined =
-			request.source === this.ctx.config.hyperbridge.config.stateMachineId
-				? RequestStatus.HYPERBRIDGE_DELIVERED
-				: RequestStatus.SOURCE
-
-		const latestMetadata = request.statuses[request.statuses.length - 1]
-		status = maxBy([status, latestMetadata.status as RequestStatusKey], (item) => REQUEST_STATUS_WEIGHTS[item])
-		if (!status) return
-
-		while (true) {
-			switch (status) {
-				case RequestStatus.SOURCE: {
-					const sourceUpdate = await waitOrAbort(this.ctx, {
-						signal,
-						promise: () =>
-							this.queries.queryStateMachineUpdateByHeight({
-								statemachineId: request.source,
-								height: request.statuses[0].metadata.blockNumber,
-								chain: this.ctx.config.hyperbridge.config.stateMachineId,
-							}),
-					})
-
-					yield {
-						status: RequestStatus.SOURCE_FINALIZED,
-						metadata: {
-							blockHash: sourceUpdate.blockHash,
-							blockNumber: sourceUpdate.height,
-							transactionHash: sourceUpdate.transactionHash,
-							timestamp: sourceUpdate.timestamp,
-						},
-					}
-					status = RequestStatus.SOURCE_FINALIZED
-					break
-				}
-
-				case RequestStatus.SOURCE_FINALIZED: {
-					request = await waitOrAbort(this.ctx, {
-						signal,
-						promise: () => this.queries.queryGetRequest(hash),
-						predicate: (r) => !r || r.statuses.length < 2,
-					})
-
-					status =
-						request.source === this.ctx.config.hyperbridge.config.stateMachineId
-							? RequestStatus.DESTINATION
-							: RequestStatus.HYPERBRIDGE_DELIVERED
-
-					yield {
-						status,
-						metadata: {
-							blockHash: request.statuses[1].metadata.blockHash,
-							blockNumber: request.statuses[1].metadata.blockNumber,
-							transactionHash: request.statuses[1].metadata.transactionHash,
-							// @ts-ignore
-							timestamp: request.statuses[1].metadata.timestamp,
-						},
-					}
-					break
-				}
-
-				case RequestStatus.HYPERBRIDGE_DELIVERED: {
-					if (request.source === this.ctx.config.hyperbridge.config.stateMachineId) return
-
-					const response = await this.queries.queryResponseByRequestId(hash)
-					yield await this.proofFinalizer.streamGetResponseFinalized(signal, request, 1, response)
-					status = RequestStatus.HYPERBRIDGE_FINALIZED
-					break
-				}
-
-				case RequestStatus.HYPERBRIDGE_FINALIZED: {
-					if (request.source === this.ctx.config.hyperbridge.config.stateMachineId) return
-
-					request = await waitOrAbort(this.ctx, {
-						signal,
-						promise: () => this.queries.queryGetRequest(hash),
-						predicate: (r) => !r || !r.statuses.find((s) => s.status === RequestStatus.DESTINATION),
-					})
-
-					yield {
-						status: RequestStatus.DESTINATION,
-						metadata: {
-							blockHash: request.statuses[2].metadata.blockHash,
-							blockNumber: request.statuses[2].metadata.blockNumber,
-							transactionHash: request.statuses[2].metadata.transactionHash,
-							// @ts-ignore
-							timestamp: request.statuses[2].metadata.timestamp,
-						},
-					}
-					status = RequestStatus.DESTINATION
-					break
-				}
-
-				case RequestStatus.DESTINATION:
-					return
-			}
-		}
-	}
-}
