+++ sdk/packages/sdk/src/storage/drivers/node.ts
@@ -1,4 +1,3 @@
-// @ts-expect-error Unable to resolve type
 import fsDriver from "unstorage/drivers/fs"
 import type { LoadDriver } from "../types"
